+++ contracts/deploy/01-outbox/01-gnosis-to-arb-outbox.ts
@@ -5,7 +5,7 @@ import getContractAddress from "../../deploy-helpers/getContractAddress";
 import { ethers } from "hardhat";
 
 enum ReceiverChains {
-  ARBITRUM_GOERLI = 421613,
+  ARBITRUM_SEPOLIA = 421614,
   ARBITRUM = 42161,
   HARDHAT = 31337,
 }
@@ -19,7 +19,7 @@ const paramsByChainId = {
     sequencerDelayLimit: 86400,
     sequencerFutureLimit: 3600,
   },
-  ARBITRUM_GOERLI: {
+  ARBITRUM_SEPOLIA: {
     deposit: parseEther("4"),
     epochPeriod: 7200, // 2 hours
     challengePeriod: 86400, // 24 hours
@@ -49,13 +49,13 @@ const deployOutbox: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
 
   const senderNetworks = {
     ARBITRUM: config.networks.gnosischain,
-    ARBITRUM_GOERLI: config.networks.chiado,
+    ARBITRUM_SEPOLIA: config.networks.chiado,
     HARDHAT: config.networks.localhost,
   };
 
   const routerNetworks = {
     ARBITRUM: config.networks.mainnet,
-    ARBITRUM_GOERLI: config.networks.goerli,
+    ARBITRUM_SEPOLIA: config.networks.sepolia,
     HARDHAT: config.networks.localhost,
   };
