+++ contracts/deploy/03-routers/03-arb-sepolia-to-chiado-router.ts
@@ -3,13 +3,13 @@ import { DeployFunction } from "hardhat-deploy/types";
 import { ethers } from "hardhat";
 
 enum RouterChains {
-  ETHEREUM_GOERLI = 5,
+  ETHEREUM_SEPOLIA = 11155111,
   HARDHAT = 31337,
 }
 
 const paramsByChainId = {
-  ETHEREUM_GOERLI: {
-    arbitrumBridge: "0xaf4159A80B6Cc41ED517DB1c453d1Ef5C2e4dB72", // https://developer.arbitrum.io/useful-addresses
+  ETHEREUM_SEPOLIA: {
+    arbitrumBridge: "0x38f918D0E9F1b721EDaA41302E399fa1B79333a9", // https://developer.arbitrum.io/useful-addresses
     amb: "0x99Ca51a3534785ED619f46A79C7Ad65Fa8d85e7a",
   },
   HARDHAT: {
@@ -21,7 +21,7 @@ const paramsByChainId = {
 // TODO: use deterministic deployments
 const deployRouter: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
   const { deployments, getNamedAccounts, getChainId } = hre;
-  const { deploy, execute } = deployments;
+  const { deploy } = deployments;
   const chainId = Number(await getChainId());
 
   // fallback to hardhat node signers on local network
@@ -32,29 +32,29 @@ const deployRouter: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
 
   // ----------------------------------------------------------------------------------------------
   const hardhatDeployer = async () => {
-    const veaOutbox = await deployments.get("VeaOutboxGnosisToArbDevnet");
-    const veaInbox = await deployments.get("VeaInboxGnosisToArbDevnet");
+    const veaOutbox = await deployments.get("VeaOutboxArbToGnosisDevnet");
+    const veaInbox = await deployments.get("VeaInboxArbToGnosisDevnet");
 
-    const router = await deploy("RouterGnosisToArbDevnet", {
+    const router = await deploy("RouterArbToGnosisDevnet", {
       from: deployer,
-      contract: "RouterGnosisToArb",
-      args: [arbitrumBridge, amb, veaInbox.address, veaOutbox.address, chainId],
+      contract: "RouterArbToGnosis",
+      args: [arbitrumBridge, amb, veaInbox.address, veaOutbox.address],
     });
   };
 
   // ----------------------------------------------------------------------------------------------
   const liveDeployer = async () => {
-    const veaOutbox = await hre.companionNetworks.arbitrumGoerli.deployments.get("VeaOutboxGnosisToArbDevnet");
-    const veaInbox = await hre.companionNetworks.chiado.deployments.get("VeaInboxGnosisToArbDevnet");
+    const veaOutbox = await hre.companionNetworks.chiado.deployments.get("VeaOutboxArbToGnosisDevnet");
+    const veaInbox = await hre.companionNetworks.arbitrumSepolia.deployments.get("VeaInboxArbToGnosisDevnet");
 
-    const router = await deploy("RouterGnosisToArbDevnet", {
+    const router = await deploy("RouterArbToGnosisDevnet", {
       from: deployer,
-      contract: "RouterGnosisToArb",
-      args: [arbitrumBridge, amb, veaInbox.address, veaOutbox.address, chainId],
+      contract: "RouterArbToGnosis",
+      args: [arbitrumBridge, amb, veaInbox.address, veaOutbox.address],
       log: true,
     });
 
-    console.log("RouterGnosisToArbDevnet deployed to: %s", router.address);
+    console.log("RouterArbToGnosisDevnet deployed to: %s", router.address);
   };
 
   // ----------------------------------------------------------------------------------------------
@@ -65,7 +65,7 @@ const deployRouter: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
   }
 };
 
-deployRouter.tags = ["ChiadoToArbGoerliRouter"];
+deployRouter.tags = ["ArbSepoliaToChiadoRouter"];
 deployRouter.skip = async ({ getChainId }) => {
   const chainId = Number(await getChainId());
   console.log(chainId);
