+++ contracts/package.json
@@ -43,21 +43,21 @@
     "deploy:devnet": "yarn deploy:ArbToEthDevnet && yarn deploy:ArbToGnosisDevnet && yarn deploy:GnosisToArbDevnet",
     "deploy:ArbToEth": "hardhat deploy --network mainnet --tags ArbToEthOutbox && hardhat deploy --network arbitrum --tags ArbToEthInbox",
     "deploy:ArbToGnosis": "hardhat deploy --network gnosis --tags ArbToGnosisOutbox && hardhat deploy --network arbitrum --tags ArbToGnosisInbox && hardhat deploy --network mainnet --tags ArbToGnosisRouter",
-    "deploy:GnosisToArb": "hardhat deploy --network arbitrum --tags GnosisToArbOutbox && hardhat deploy --network chiado --tags GnosisToArbInbox && hardhat deploy --network goerli --tags GnosisToArbRouter",
-    "deploy:ArbToEthTestnet": "hardhat deploy --network goerli --tags ArbToEthOutbox && hardhat deploy --network arbitrumGoerli --tags ArbToEthInbox",
-    "deploy:ArbToGnosisTestnet": "hardhat deploy --network chiado --tags ArbToGnosisOutbox && hardhat deploy --network arbitrumGoerli --tags ArbToGnosisInbox && hardhat deploy --network goerli --tags ArbToGnosisRouter",
-    "deploy:GnosisToArbTestnet": "hardhat deploy --network arbitrumGoerli --tags GnosisToArbOutbox && hardhat deploy --network chiado --tags GnosisToArbInbox && hardhat deploy --network goerli --tags GnosisToArbRouter",
-    "deploy:ArbToEthDevnet": "hardhat deploy --network goerli --tags ArbGoerliToGoerliOutbox && hardhat deploy --network arbitrumGoerli --tags ArbGoerliToGoerliInbox",
-    "deploy:ArbToGnosisDevnet": "hardhat deploy --network chiado --tags ArbGoerliToChiadoOutbox && hardhat deploy --network arbitrumGoerli --tags ArbGoerliToChiadoInbox && hardhat deploy --network goerli --tags ArbGoerliToChiadoRouter",
-    "deploy:GnosisToArbDevnet": "hardhat deploy --network arbitrumGoerli --tags ChiadoToArbGoerliOutbox && hardhat deploy --network chiado --tags ChiadoToArbGoerliInbox && hardhat deploy --network goerli --tags ChiadoToArbGoerliRouter",
+    "deploy:GnosisToArb": "hardhat deploy --network arbitrum --tags GnosisToArbOutbox && hardhat deploy --network chiado --tags GnosisToArbInbox && hardhat deploy --network sepolia --tags GnosisToArbRouter",
+    "deploy:ArbToEthTestnet": "hardhat deploy --network sepolia --tags ArbToEthOutbox && hardhat deploy --network arbitrumSepolia --tags ArbToEthInbox",
+    "deploy:ArbToGnosisTestnet": "hardhat deploy --network chiado --tags ArbToGnosisOutbox && hardhat deploy --network arbitrumSepolia --tags ArbToGnosisInbox && hardhat deploy --network sepolia --tags ArbToGnosisRouter",
+    "deploy:GnosisToArbTestnet": "hardhat deploy --network arbitrumSepolia --tags GnosisToArbOutbox && hardhat deploy --network chiado --tags GnosisToArbInbox && hardhat deploy --network sepolia --tags GnosisToArbRouter",
+    "deploy:ArbToEthDevnet": "hardhat deploy --network sepolia --tags ArbSepoliaToSepoliaOutbox && hardhat deploy --network arbitrumSepolia --tags ArbSepoliaToSepoliaInbox",
+    "deploy:ArbToGnosisDevnet": "hardhat deploy --network chiado --tags ArbSepoliaToChiadoOutbox && hardhat deploy --network arbitrumSepolia --tags ArbSepoliaToChiadoInbox && hardhat deploy --network sepolia --tags ArbSepoliaToChiadoRouter",
+    "deploy:GnosisToArbDevnet": "hardhat deploy --network arbitrumSepolia --tags ChiadoToArbSepoliaOutbox && hardhat deploy --network chiado --tags ChiadoToArbSepoliaInbox && hardhat deploy --network sepolia --tags ChiadoToArbSepoliaRouter",
     "deploy-local": "hardhat deploy --tags ArbToEthOutbox,ArbToEthInbox --network localhost",
-    "etherscan-verify:testnet": "yarn etherscan-verify:chiado && yarn etherscan-verify:goerli && yarn etherscan-verify:arbitrumGoerli",
+    "etherscan-verify:testnet": "yarn etherscan-verify:chiado && yarn etherscan-verify:sepolia && yarn etherscan-verify:arbitrumSepolia",
     "etherscan-verify:gnosis": "hardhat etherscan-verify --network gnosischain",
     "etherscan-verify:chiado": "hardhat etherscan-verify --network chiado",
     "etherscan-verify:ethereum": "hardhat etherscan-verify --network ethereum",
-    "etherscan-verify:goerli": "hardhat etherscan-verify --network goerli",
+    "etherscan-verify:sepolia": "hardhat etherscan-verify --network sepolia",
     "etherscan-verify:arbitrum": "hardhat etherscan-verify --network arbitrum",
-    "etherscan-verify:arbitrumGoerli": "hardhat etherscan-verify --network arbitrumGoerli",
+    "etherscan-verify:arbitrumSepolia": "hardhat etherscan-verify --network arbitrumSepolia",
     "sourcify": "hardhat sourcify --write-failing-metadata",
     "size": "hardhat size-contracts --no-compile",
     "watch": "hardhat watch",
